refs #30061 MA phasequad clean up
