+++ scripts/Muon/GUI/Common/grouping_tab_widget/grouping_tab_widget_model.py
@@ -113,6 +113,7 @@ class GroupingTabModel(object):
         self._groups_and_pairs.add_group(group)
 
     def add_pair(self, pair):
+        print(":P")
         assert isinstance(pair, MuonPair)
         self._groups_and_pairs.add_pair(pair)
