+++ scripts/Muon/GUI/Common/phase_table_widget/phase_table_view.py
@@ -70,16 +70,6 @@ class PhaseTableView(QtWidgets.QWidget, ui_muon_phases_tab):
             self.backward_group_combo.setCurrentIndex(index)
 
     @property
-    def phase_quad_input_workspace(self):
-        return str(self.phase_quad_input_workspace_combo.currentText())
-
-    @phase_quad_input_workspace.setter
-    def phase_quad_input_workspace(self, value):
-        index = self.phase_quad_input_workspace_combo.findText(value)
-        if index != -1:
-            self.phase_quad_input_workspace_combo.setCurrentIndex(index)
-
-    @property
     def phase_table_for_phase_quad(self):
         return str(self.phase_quad_phase_table_combo.currentText())
 
@@ -97,10 +87,6 @@ class PhaseTableView(QtWidgets.QWidget, ui_muon_phases_tab):
         self.input_workspace_combo_box.clear()
         self.input_workspace_combo_box.addItems(input_list)
         self.input_workspace_combo_box.setCurrentIndex(0)
-
-        self.phase_quad_input_workspace_combo.clear()
-        self.phase_quad_input_workspace_combo.addItems(input_list)
-        self.phase_quad_input_workspace_combo.setCurrentIndex(0)
 
     def set_group_combo_boxes(self, group_list):
         self.forward_group_combo.clear()
@@ -206,14 +192,9 @@ class PhaseTableView(QtWidgets.QWidget, ui_muon_phases_tab):
         table_utils.setTableHeaders(self.phase_table_options_table)
 
         self.phase_quad_table.setColumnWidth(0, 300)
-        self.phase_quad_table.setColumnWidth(1, 300)
 
-        # self.calculate_phase_table_button = QtWidgets.QPushButton('Calculate Phase Table', self)
-        table_utils.setRowName(self.phase_quad_table, 0, 'InputWorkspace')
-        self.phase_quad_input_workspace_combo = table_utils.addComboToTable(self.phase_quad_table, 0, options)
-
-        table_utils.setRowName(self.phase_quad_table, 1, 'PhaseTable')
-        self.phase_quad_phase_table_combo = table_utils.addComboToTable(self.phase_quad_table, 1, options)
+        table_utils.setRowName(self.phase_quad_table, 0, 'PhaseTable')
+        self.phase_quad_phase_table_combo = table_utils.addComboToTable(self.phase_quad_table, 0, options)
 
         self.phase_quad_table.resizeRowsToContents()
